Fix bug in shifting UByte/UShort + clean up

+++ core/src/main/scala/spire/math/Natural.scala
@@ -36,7 +36,7 @@ sealed trait Natural {
 
   def getNumBits: Int = {
     @tailrec
-    def bit(n: UInt, b: Int): Int = if (n == 0) b else bit(n >>> UInt(1), b + 1)
+    def bit(n: UInt, b: Int): Int = if (n == 0) b else bit(n >>> 1, b + 1)
 
     @tailrec
     def recur(next: Natural, b: Int): Int = next match {
@@ -154,7 +154,7 @@ sealed trait Natural {
       if ((n.signed & -n.signed) != n.signed) return -1
       // TODO: this could be better/faster
       var i = 1
-      while (i < 32 && (n >>> UInt(i)) != 0) i += 1
+      while (i < 32 && (n >>> i) != 0) i += 1
       i - 1
     }
 
@@ -327,8 +327,8 @@ sealed trait Natural {
   def pow(rhs: UInt): Natural = {
     @tailrec def _pow(t: Natural, b: Natural, e: UInt): Natural = {
       if (e == UInt(0)) t
-      else if ((e & UInt(1)) == UInt(1)) _pow(t * b, b * b, e >> UInt(1))
-      else _pow(t, b * b, e >> UInt(1))
+      else if ((e & UInt(1)) == UInt(1)) _pow(t * b, b * b, e >> 1)
+      else _pow(t, b * b, e >> 1)
     }
     _pow(Natural(1), lhs, rhs)
   }
